Fix GitLab CI issues of ant.
Specify ivy cache in build directory so that
caching works, and introduce a new ant
property `skipBuild`, which allows us to tell ant
to always skip the build for the testing stage.
(The testing stage receives the build from the previous stage.
We have to skip the build because the wrong order of timestamps
in newly cloned directories makes `ant build` recompile the whole
project.)