Don't dispose of Memory object through incompatible LLVM method.
Fixes a double-free error.
The method `LLVMDisposeMessage(Pointer)` disposes of the given memory,
but it doesn't set the Java field of the corresponding JNA Memory object to 0.
JNA Memory objects dispose of the underlying memory in their finalizer
if the pointer is not set to 0, thus resulting in an error upon garbage
collection.