Use libLLVM-6.0 from Ubuntu 18.04 instead of 20.04
The libLLVM-6.0 version of Ubuntu 20.04
is compiled against glibc 2.29, which is not available
for Ubuntu 18.04. Since Ubuntu 18.04 is still supported
until 2023, it may be reckless to only support 20.04.

As a downside, both the official libLLVM-6.0 package provided
by LLVM and the one provided by the Ubuntu repository are compiled
without support for LLVMDump*, so we are not able to dump values
anymore until we upgrade to the libraries in 20.04.
While nice for debugging, LLVMDump* was not used otherwise,
so it's a valid tradeoff.

On a minor note, LLVMInitializeInstCombine is also not supported,
but we didn't provide an interface for that yet.